add Sphere shape; use spherical light sources; implement per-vertex lighting and sphere normals; add scene-managed Orbit components; refactor Plane and Cube vertex storage; preserve projection aspect ratio; clean up rendering code and comments

+++ src/SFMLApp/Shapes/Base/Shape3D.cs
@@ -20,7 +20,7 @@ public abstract class Shape3D
     /// <summary>
     /// Represents vertices
     /// </summary>
-    protected int[][] _model { get; set; } = null!;
+    protected float[][] _model { get; set; } = null!;
 
     public abstract IEnumerable<DrawCall> CollectFaces(Camera camera, IReadOnlyList<LightSource> lightSources);
 }
